+++ Package.swift
@@ -17,12 +17,21 @@ let package = Package(
             dependencies: [],
             path: "Sources/Firelink",
             exclude: [
-                "_internal"
+                "deno-version.txt",
+                "ffmpeg-version.txt"
             ],
             resources: [
                 .process("Assets.xcassets"),
                 .copy("yt-dlp"),
-                .copy("ffmpeg")
+                .copy("yt-dlp-version.txt"),
+                .copy("_internal"),
+                .copy("deno"),
+                .copy("ffmpeg"),
+                .copy("aria2c"),
+                .copy("aria2-libs"),
+                .copy("aria2-cacert.pem"),
+                .copy("aria2-version.txt"),
+                .copy("aria2-licenses")
             ]
         )
     ]